Achieved good basis for extendability
