+++ ena_portal_api/query/operators.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+class ENAQueryOperators(str, Enum):
+    OR = "OR"
+    AND = "AND"
+    NOT = "NOT"
