+++ requirements.txt
@@ -1,1 +1,6 @@
 requests==2.31.0
+
+setuptools~=75.6.0
+pydantic~=2.10.6
+urllib3~=2.3.0
+httpx~=0.28.1
